update on customer

+++ internal/service/unlink/core/helper.go
@@ -8,7 +8,7 @@ import (
 	"context"
 )
 
-func CreatArchiveUserDataWithLinkedAccount(ctx context.Context, archivedUserRepo storage.ArchivedUserRepository, archivedLinkedAccountRepo storage.ArchivedLinkedAccountRepository, userOldData *model.User, linkedAccountOldData *model.LinkedAccount) (error, error) {
+func CreatArchiveUserDataWithLinkedAccount(ctx context.Context, archivedUserRepo storage.ArchivedUserRepository, archivedLinkedAccountRepo storage.ArchivedLinkedAccountRepository, userOldData *model.User, linkedAccountOldData *model.LinkedAccount, haveAccount bool) (error, error) {
 	var archUserErr, archLinkedAccErr error
 
 	lib.GoRoutinBaker(types.BakerOptions{Sequential: false, UseMutex: false},
@@ -16,14 +16,16 @@ func CreatArchiveUserDataWithLinkedAccount(ctx context.Context, archivedUserRepo
 			archUserErr = archivedUserRepo.Create(ctx, userOldData)
 		},
 		func() {
-			archLinkedAccErr = archivedLinkedAccountRepo.Create(ctx, linkedAccountOldData)
+			if haveAccount {
+				archLinkedAccErr = archivedLinkedAccountRepo.Create(ctx, linkedAccountOldData)
+			}
 		},
 	)
 
 	return archUserErr, archLinkedAccErr
 }
 
-func DeleteUserDataWithLinkedAccount(ctx context.Context, userRepo storage.UserRepository, LinkedAccountRepo storage.LinkedAccountRepository, userID string, likedAccountId string) (error, error) {
+func DeleteUserDataWithLinkedAccount(ctx context.Context, userRepo storage.UserRepository, LinkedAccountRepo storage.LinkedAccountRepository, userID string, likedAccountId string, haveAccount bool) (error, error) {
 	var archUserErr, archLinkedAccErr error
 	// archUserErr = userRepo.Delete(ctx, userID)
 	// archLinkedAccErr = LinkedAccountRepo.Delete(ctx, likedAccountId)
@@ -33,7 +35,9 @@ func DeleteUserDataWithLinkedAccount(ctx context.Context, userRepo storage.UserR
 			archUserErr = userRepo.Delete(ctx, userID)
 		},
 		func() {
-			archLinkedAccErr = LinkedAccountRepo.Delete(ctx, likedAccountId)
+			if haveAccount {
+				archLinkedAccErr = LinkedAccountRepo.Delete(ctx, likedAccountId)
+			}
 		},
 	)
